v2.4.4.1 -> v2.4.4.2
  - David Miller: network syncup (better TCP page fault handling by Andrea etc)
  - Tim Hockin: moxa driver fixes
  - Alexander Viro: clean up process reaping procedures, make /proc use
  the proper fields to notice that the process is gone, gone, gone.
  - Justin Gibbs: aic7xxx driver update to 6.1.13.
  - me: fix PCI allocation so that it doesn't stomp on unknown memory
  - Jeff Garzik: 8139too net drvr fix
  - Nigel Gamble: merge page fault interrupt gate from preemtible kernel.
  Fixes problem noticed by Brian Gerst.
  - Jean Tourrilhes: Orinoco / Airport support diff. Mucho cleanups.

+++ drivers/net/wireless/hermes.c
@@ -106,7 +106,7 @@ static int hermes_issue_cmd(hermes_t *hw, uint16_t cmd, uint16_t param0)
  * Function definitions
  */
 
-void hermes_struct_init(hermes_t *hw, ushort io)
+void hermes_struct_init(hermes_t *hw, uint io)
 {
 	hw->iobase = io;
 	hw->inten = 0x0;
